+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/usgs/waterdataui/compare/waterdataui-0.20.0...master)
+## [Unreleased](https://github.com/usgs/waterdataui/compare/waterdataui-0.21.0...master)
+
+## [0.21.0](https://github.com/usgs/waterdataui/compare/waterdataui-0.20.0...waterdataui-0.21.0) - 2019-10-24
 ### Added
 - Added a checkbox to enable/disable median line on hydrograph